requirements.txt
python-telegram-bot==20.7
requests==2.32.3
beautifulsoup4==4.12.3
lxml==4.9.4
sqlalchemy==2.0.27
psycopg[binary]>=3.1,<4
python-dotenv==1.0.1
openpyxl==3.1.2
pandas==2.2.0
aiohttp==3.10.0
